asdsaf

+++ MatAppen/Food.swift
@@ -13,7 +13,7 @@ class Food {
     
     var name: String
     var id: Int
-    var kcal: Double
+    var kcal: Int
     var protein: Double
     var fat: Double
     var carbs: Double
@@ -22,7 +22,7 @@ class Food {
     let defaults = UserDefaults.standard
     var healthiness = 0.0
     
-    init(name: String = "", id: Int = 0, kcal: Double = 0.0,protein: Double = 0.0, fat: Double = 0.0, carbs: Double = 0.0, healthiness: Double = 0.0) {
+    init(name: String = "", id: Int = 0, kcal: Int = 0,protein: Double = 0.0, fat: Double = 0.0, carbs: Double = 0.0, healthiness: Double = 0.0) {
         self.name = name
         self.id = id
         self.kcal = kcal
@@ -31,7 +31,7 @@ class Food {
         self.carbs = carbs
         self.healthiness = healthiness
     }
-    func isFavorite(name: String, id: Int, kcal: Double,protein: Double, fat: Double, carbs: Double, isFav: Bool) {
+    func isFavorite(name: String, id: Int, kcal: Int,protein: Double, fat: Double, carbs: Double, isFav: Bool) {
         if let favorites = defaults.array(forKey: "favorites") as? [Dictionary<String, Any>] {
             favArray = favorites
